서버: 루아 객체 제거,  IOCP 클래스 IOCP 핸들 리턴함수 추가, EVENT_TIMER와 IOCP연동을 위한 테스트 case 추가(미 테스트), 로직 클래스 roomId를 이용한 패킷 broadCast함수 생성
프리 컴파일 헤더에 Timer_Event구조체 생성, IOCP 명렁어 enum 위치 변경, 몬스터 섹션 클래스 루아관련 함수 삭제, 타이머 클래스 생성 및 쓰레드 돌면서 이벤트 처리할 수 있게 추가

+++ GRADUATION PROJECT/MyFrameWork/Server/Session/SessionObject/PlayerSessionObject.cpp
@@ -122,7 +122,7 @@ void PlayerSessionObject::AutoMove()
 	//	}
 	//}
 	// m_position = Vector3::Add(m_position, Vector3::ScalarProduct(m_directionVector, ((double)durationTime / 1000.0f) * 50.0f));
-	Move(((double)durationTime / 1000.0f) * 50.0f);
+	Move(((float)durationTime / 1000.0f) * 50.0f);
 	m_lastMoveTime = currentTime;
 	std::cout << "current Position " << m_position.x << " " << m_position.y << " " << m_position.z << std::endl;
 	//std::cout << "rotate angle" << m_rotateAngle.x << " " << m_rotateAngle.y << " " << m_rotateAngle.z << std::endl;
@@ -350,7 +350,7 @@ void PlayerSessionObject::SetDirection(DIRECTION d)
 		case DIRECTION::FRONT | DIRECTION::RIGHT:	fRotateAngle = 315.0f;	break;
 		}
 
-		fRotateAngle = fRotateAngle * (3.14159265359 / 180.0f);
+		fRotateAngle = fRotateAngle * (3.14159265359f / 180.0f);
 		xmf3Rev.x = xmf3Look.x * cos(fRotateAngle) - xmf3Look.z * sin(fRotateAngle);
 		xmf3Rev.z = xmf3Look.x * sin(fRotateAngle) + xmf3Look.z * cos(fRotateAngle);
 		xmf3Rev = Vector3::Normalize(xmf3Rev);
